+++ astropy/cosmology/core.py
@@ -263,6 +263,13 @@ class FLRW(Cosmology):
             if self._massivenu:
                 nu_y = self._massivenu_mass / (kB_evK * self._Tnu0)
                 self._nu_y = nu_y.value
+                # This may seem strange, but it turns out to
+                # be much faster to use _nu_y as a python list than
+                # an ndarray for scalar evaluations; see
+                # _massivenu_relative_density_scalar.  So we provide
+                # both -- the ndarray for when z is an ndarray, and
+                # the list version for inside the integrals
+                self._nu_ylist = self._nu_y.tolist()
                 self._Onu0 = self._Ogamma0 * self.nu_relative_density(0)
             else:
                 # This case is particularly simple, so do it directly
@@ -749,7 +756,14 @@ class FLRW(Cosmology):
         analytical fitting formula given in Komatsu et al. 2011, ApJS 192, 18.
         """
 
-        # See Komatsu et al. 2011, eq 26 and the surrounding discussion
+        # This is the vectorized version meant for direct use.
+        # It is substantially more efficient to also have a scalar-only
+        # version (since that is what the distance, etc., integrals
+        # end up calling); see FLRW._get_nufunc_scalar and
+        # FLRW._massivenu_relative_density_scalar.
+
+        # For the mathematical details, see Komatsu et al. 2011, eq 26 and
+        # the surrounding discussion.
         # However, this is modified to handle multiple neutrino masses
         # by computing the above for each mass, then summing
         prefac = 0.22710731766  # 7/8 (4/11)^4/3 -- see any cosmo book
@@ -766,8 +780,8 @@ class FLRW(Cosmology):
         p = 1.83
         invp = 1.0 / p
         if np.isscalar(z):
-            curr_nu_y = self._nu_y / (1.0 + z)  # only includes massive ones
-            rel_mass_per = (1.0 + (0.3173 * curr_nu_y) ** p) ** invp
+            k = 0.3173 / (1.0 + z)
+            rel_mass_per = (1.0 + (k * self._nu_y) ** p) ** invp
             rel_mass = rel_mass_per.sum() + self._nmasslessnu
         else:
             z = np.asarray(z)
@@ -778,113 +792,17 @@ class FLRW(Cosmology):
         return prefac * self._neff_per_nu * rel_mass
 
     def _get_nufunc_scalar(self):
-        """ Get the scalar nu density function for integrals"""
+        """ Return the version of the nu relative density function
+        to use in integrals."""
 
-        nu_dict = {0: lambda x : 0.22710731766 * self._Neff,
-                   1: self._nu_relative_density_scalar_single,
-                   2: self._nu_relative_density_scalar_double,
-                   3: self._nu_relative_density_scalar_triple}
-        return nu_dict.get(self._nmassivenu,
-                           self._nu_relative_density_scalar)
+        if self._massivenu:
+            return self._massivenu_relative_density_scalar
+        else:
+            return lambda x : 0.22710731766 * self._Neff
 
-    def _nu_relative_density_scalar_single(self, z):
+    def _massivenu_relative_density_scalar(self, z):
         """ Neutrino density function relative to the energy density in
-        photons.
-
-        Parameters
-        ----------
-        z : scalar
-           Redshift
-
-        Returns
-        -------
-         f : float
-           The neutrino density scaling factor relative to the density
-           in photons at the specified redshift
-
-        Notes
-        -----
-        This version is specialized to scalar z and a single
-        massive neutrino flavor for performance reasons.
-        """
-
-        prefac = 0.22710731766
-        p = 1.83
-        invp = 0.5464480874316939
-        curr_nu_y = self._nu_y[0] / (1.0 + z)
-        rel_mass_per = (1.0 + (0.3173 * curr_nu_y) ** p) ** invp
-        rel_mass = rel_mass_per + self._nmasslessnu
-        return prefac * self._neff_per_nu * rel_mass
-
-    def _nu_relative_density_scalar_double(self, z):
-        """ Neutrino density function relative to the energy density in
-        photons.
-
-        Parameters
-        ----------
-        z : scalar
-           Redshift
-
-        Returns
-        -------
-         f : float
-           The neutrino density scaling factor relative to the density
-           in photons at the specified redshift
-
-        Notes
-        -----
-        This version is specialized to scalar z and two
-        massive neutrino flavor for performance reasons.
-        """
-
-        prefac = 0.22710731766
-        p = 1.83
-        invp = 0.5464480874316939
-        invz = 1.0 / (1.0 + z)
-        curr_nu_y = self._nu_y[0] * invz
-        rel_mass_per = (1.0 + (0.3173 * curr_nu_y) ** p) ** invp
-        curr_nu_y = self._nu_y[1] * invz
-        rel_mass_per += (1.0 + (0.3173 * curr_nu_y) ** p) ** invp
-        rel_mass = rel_mass_per + self._nmasslessnu
-        return prefac * self._neff_per_nu * rel_mass
-
-    def _nu_relative_density_scalar_triple(self, z):
-        """ Neutrino density function relative to the energy density in
-        photons.
-
-        Parameters
-        ----------
-        z : scalar
-           Redshift
-
-        Returns
-        -------
-         f : float
-           The neutrino density scaling factor relative to the density
-           in photons at the specified redshift
-
-        Notes
-        -----
-        This version is specialized to scalar z and three
-        massive neutrino flavor for performance reasons.
-        """
-
-        prefac = 0.22710731766
-        p = 1.83
-        invp = 0.5464480874316939
-        invz = 1.0 / (1.0 + z)
-        curr_nu_y = self._nu_y[0] * invz
-        rel_mass_per = (1.0 + (0.3173 * curr_nu_y) ** p) ** invp
-        curr_nu_y = self._nu_y[1] * invz
-        rel_mass_per += (1.0 + (0.3173 * curr_nu_y) ** p) ** invp
-        curr_nu_y = self._nu_y[2] * invz
-        rel_mass_per += (1.0 + (0.3173 * curr_nu_y) ** p) ** invp
-        rel_mass = rel_mass_per + self._nmasslessnu
-        return prefac * self._neff_per_nu * rel_mass
-
-    def _nu_relative_density_scalar(self, z):
-        """ Neutrino density function relative to the energy density in
-        photons.
+        photons for massive neutrinos.
 
         Parameters
         ----------
@@ -900,15 +818,18 @@ class FLRW(Cosmology):
         Notes
         -----
         This version is specialized to scalar z for performance reasons.
+        It also assumes there are actually massive neutrinos present.
         """
 
-        prefac = 0.22710731766
-        p = 1.83
-        invp = 0.5464480874316939
-        curr_nu_y = self._nu_y / (1.0 + z)
-        rel_mass_per = (1.0 + (0.3173 * curr_nu_y) ** p) ** invp
-        rel_mass = rel_mass_per.sum() + self._nmasslessnu
-        return prefac * self._neff_per_nu * rel_mass
+        # See nu_relative_density for an explanation of the formula
+        # Bizarrely(?), it's -much- faster if _nu_y is available as a
+        # python list instead of a numpy ndarray manipulated using
+        # np.sum, etc.
+        rel_mass = float(self._nmasslessnu)
+        k = 0.3173 / (1.0 + z)
+        for curr_nu_y in self._nu_ylist:
+            rel_mass += ((1.0 + (k * curr_nu_y) ** 1.83)**0.5464480874316939)
+        return 0.22710731766 * self._neff_per_nu * rel_mass
 
     def _w_integrand(self, ln1pz):
         """ Internal convenience function for w(z) integral."""